Refactor. Now each page separated by folder, more organized components
